Fix for bug 1196140 . This bug prevented running Wicket on VM's that don't use SUN's security implementation classes. Don Ledford reported it for Blackdown 1.4.2 j2sdk.
Removed SunJceCrypt's static initializer and put it in it's constructor. Application now reuses the ICrypt instance once it is created, so that it is easier to provide a dummy one. The current ICrypt instances seem to be thread safe, so that shouldn't give any problems, and users can override the factory method if they want to use non-thread safe instances.


git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/wicket/trunk@456659 13f79535-47bb-0310-9956-ffa450edef68