rfe 1349466.
Excellent suggestion. I made a couple of changes though, to
bridge with the ajaxhandler more nicely:

* I made it an interface (IBehaviour) instead of an abstract
class. In this case, there was no reason (no behaviour) for
an abstract class so an interface is more flexible for users.

* AbstractBehaviour serves as an adapter/ empty
implementation so that users can more easily create
annonymous class implementations etc.

* IBehaviourRequest is now what previously was
IAjaxListener; it serves the same purpose with the
difference that it is now fine for users to implement this
interface directly (contrairy to IAjaxListener).

* IBehaviour extends interfaces IHeaderContributor and
IBodyOnloadContributor so that they can make contributions
for javascript etc. This is needed for ajax implementations
anyway, but could certainly come in handy for a lot of
custom behaviours.


git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/wicket/trunk@458017 13f79535-47bb-0310-9956-ffa450edef68