PageMap refactor. The default base pagemap doesn't do much anymore What was our previoud pagemap is now called AccessStackPageMap
this was needed for:

ClientPageSavingSessionStore A tryout of clientside page state saving. It works currently only for forms. Links are not rewritten yet to use a form submit.

SecondLevelCacheSessionStore This is for now the default one. This only holds one page (the last/current one) for one pagemap and the rest is first cached
in a soft cache store. And written to the work dir on the HD.


IConverter/ITypConverter changes.

Now we only have one IConverter with 2 methods: convertToObject/convertToString.
And a converter for a specific type can be get in the new refactored method getConverted(Class) on component or session.
This method comes from the ISupplyConverter interface which has a default ConverterSupplier class
where the sessions gets its type converters from. But you can make youre own.
It is now much easier to create converters for a component..

added Form.setBookmarkableHint(boolean), If this is set and the Page is stateless and the Page has a default or pageparams constructor
the form will generate a bookmarkable url to itself.  So the page is still stateless..
This will result ofcourse in lost page state. Because when such an url is hit. The page will first be created. then the form onFormSubmit() will be called.
So the form will always be in the constructed state. The nice thing is. That everythign just works. required chec, Type conversion and Validation.
And if you only care about the input in the onsubmit you won't notice any difference.

Large parts are java5 generified  now. This is by far done. But i got as much as i came alonge when walking through the code.

removed wicket.concurrent package now. Because you really should use java5 for that.


git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/wicket/trunk@460642 13f79535-47bb-0310-9956-ffa450edef68