making updatemodel() nonfinal. sometimes it is nice to be able to force a model.setobject() which components that work with collections do not call.
my scenario is this: this component works with a list of objects, but in session i need to store a list of ids. i have a model that converts list of ids to list of objects by loading them from db, and viceversa. in order for this model to work both get and set object methods must be called. this cannot be done with a converter, so overriding updatemodel() and calling { super(); getModel().setObject(getModelObject()); } is the best i can come up with for now.

any objections?

git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/wicket/branches/wicket-1.x@462434 13f79535-47bb-0310-9956-ffa450edef68