Fixed a couple of test failures (and bugs).
Yesterday, I sent this http://www.nabble.com/not-happy-with-how-we-did-the-backports-tf3534134.html, where I state I'm not happy with amongst
other things:
* removal of methods from the API without providing a final method to
help upgrading. If a class has overridable method fooXX(Blah b) and
that is removed (e.g. because it is renamed to barXX or got another
signature), we should make fooXX(Blah) final for at least the first
release, with a TODO to remove it later. 

This happened again, and *again* I'm fixing other's mess. It's not like I doing that for fun, nor do I have lots of free time. Please everyone,
take API changes seriously and make the methods/ signatures that are replaced final. Thanks.

git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/wicket/branches/wicket-1.x@526269 13f79535-47bb-0310-9956-ffa450edef68