WICKET-698 now for a list i do try to find a field when the rest is failing:
findMethod if expresion is () or findGetter
if method not found then look if it is an index.
if that fails then it will look for a method anyway and last it will look for the field

For a Map this doesnt happen. You can't have fields in a map you can only target
fields through a getter in a Map or explicitly by defining map.xxx() so it will just look up the method. 

git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/wicket/trunk@550777 13f79535-47bb-0310-9956-ffa450edef68