wicket-1224: PopupCloseLink broken with 1.3.0-rc2 in Internet Explorer
It was changed as part of wicket-1139 but IE doesn't like it. In order to avoid the problem I usually add to my apps
getMarkupSettings().setStripXmlDeclarationFromOutput(true);

and

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">

I wonder if especially setStripXmlDeclarationFromOutput(true) wouldn't be a better default due to IE problems. The http header still contains the encoding.

I leave this case open until we found a fix for both problem, but remove the xml declaration for the file to work properly


git-svn-id: https://svn.eu.apache.org/repos/asf/wicket/trunk@606968 13f79535-47bb-0310-9956-ffa450edef68